backend/requirements.txt
Django>=3.1.0, <3.2.0
djangorestframework >= 3.12.0, < 3.13.0
graphene-django==2.13.0
django-graphql-jwt==0.3.1
PyJWT==1.7.1
psycopg2-binary
django-environ
django-debug-toolbar
django-jazzmin
dj-rest-auth[with_social]
django-allauth
djangorestframework-simplejwt
django-cors-headers
requests_oauthlib
django-cleanup
python-magic
freezegun
django-filter

requests
beautifulsoup4

flake8 >= 3.8.0, < 3.9.0
